Changed behavior of missing type property
Created a utilities file to store common utility functions.
