+++ tests/generate_cases/empty.json
@@ -1,0 +1,1 @@
+{}
